+++ adapters/adserver_adtarget/adserver_adtarget_test/exemplary/banner-web.json
