+++ adapters/adserver_adtarget/adserver_adtarget_test/exemplary/native-app.json
