+++ adapters/adserver_adtarget/adserver_adtarget_test/exemplary/native-web.json
