+++ adapters/adserver_adtarget/adserver_adtarget_test/exemplary/video-app.json
