+++ adapters/adserver_adtarget/adserver_adtarget_test/supplemental/empty-seatbid-array.json
