+++ exchange/adapter_builders.go
@@ -19,8 +19,8 @@ import (
 	"github.com/prebid/prebid-server/adapters/adot"
 	"github.com/prebid/prebid-server/adapters/adpone"
 	"github.com/prebid/prebid-server/adapters/adprime"
+	"github.com/prebid/prebid-server/adapters/adserver_adtarget"
 	"github.com/prebid/prebid-server/adapters/adtarget"
-	"github.com/prebid/prebid-server/adapters/adtargetme"
 	"github.com/prebid/prebid-server/adapters/adtelligent"
 	"github.com/prebid/prebid-server/adapters/advangelists"
 	"github.com/prebid/prebid-server/adapters/adview"
@@ -168,7 +168,7 @@ func newAdapterBuilders() map[openrtb_ext.BidderName]adapters.Builder {
 		openrtb_ext.BidderAdot:              adot.Builder,
 		openrtb_ext.BidderAdprime:           adprime.Builder,
 		openrtb_ext.BidderAdtarget:          adtarget.Builder,
-		openrtb_ext.BidderAdtargetme:        adtargetme.Builder,
+		openrtb_ext.BidderAdserverAdtarget:  adserver_adtarget.Builder,
 		openrtb_ext.BidderAdtelligent:       adtelligent.Builder,
 		openrtb_ext.BidderAdvangelists:      advangelists.Builder,
 		openrtb_ext.BidderAdView:            adview.Builder,
