+++ static/bidder-info/adserver_adtarget.yaml
