--- static/bidder-params/adtargetme.json
@@ -1,8 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-04/schema#",
-  "title": "Adtargetme Adapter Params",
-  "description": "A schema which validates params accepted by the Adtargetme adapter",
-  "type": "object",
-
-  "properties": {}
-}
